+++ wasm-runtime/polkadot/src/runtime/consensus.rs
@@ -2,6 +2,7 @@ use runtime_support::Vec;
 use keyedvec::KeyedVec;
 use storage::Storage;
 use primitives::{AccountID, SessionKey, BlockNumber};
+use runtime::{system, staking};
 
 pub fn set_authority(index: u32, authority: AccountID) {
 	authority.store(&index.to_keyed_vec(b"con\0aut\0"));
@@ -31,42 +32,4 @@ pub fn authorities() -> Vec<AccountID> {
 pub fn set_authorities(authorities: &[AccountID]) {
 	set_authority_count(authorities.len() as u32);
 	authorities.iter().enumerate().for_each(|(v, &i)| set_authority(v as u32, i));
-}
-
-/// Get the current set of validators. These are the long-term identifiers for the validators
-/// and will be mapped to a session key with the most recent `set_next_session_key`.
-pub fn validators() -> Vec<AccountID> {
-	unimplemented!()
-}
-
-/// Set the current set of validators.
-///
-/// Called by staking::next_era() only.
-pub fn set_validators(_new: &[AccountID]) {
-	unimplemented!()
-}
-
-/// The number of blocks in each session.
-pub fn session_length() -> BlockNumber {
-	Storage::into(b"con\0bps")
-}
-
-/// Sets the session key of `_transactor` to `_session`. This doesn't take effect until the next
-/// session.
-pub fn set_session_key(_transactor: &AccountID, _session: &AccountID) {
-	unimplemented!()
-}
-
-/// Move onto next session: register the new authority set.
-pub fn next_session() {
-	// TODO: Call set_authorities().
-	unimplemented!()
-}
-
-/// Hook to be called prior to transaction processing.
-pub fn pre_transactions() {}
-
-/// Hook to be called after to transaction processing.
-pub fn post_transactions() {
-	// TODO: check block number and call next_session if necessary.
 }
